+++ src/sns/report/neurons_cache/attempt/mod.rs
@@ -4,7 +4,10 @@ mod read;
 mod timestamp;
 mod write;
 
-pub(super) use failure::failed_attempt_from_latest_progress;
-pub(super) use model::{SnsNeuronsAttemptParts, attempt_from_parts};
+pub(super) use failure::write_failed_sns_neurons_attempt;
+pub(super) use model::{SnsNeuronsAttemptContext, SnsNeuronsAttemptProgress};
 pub(super) use read::read_sns_neurons_attempt_status;
-pub(super) use write::write_sns_neurons_attempt;
+pub(super) use write::{
+    write_complete_sns_neurons_attempt, write_running_sns_neurons_attempt,
+    write_starting_sns_neurons_attempt,
+};